Restructure to extract common things, and separate armeria-specific things.
Also add an editorconfig to get intellij's auto formatting to simulate ktfmt.
