+++ api/src/main/kotlin/com/geekinasuit/daggergrpc/api/GrpcServiceHandler.kt
@@ -1,0 +1,6 @@
+package com.geekinasuit.daggergrpc.api
+
+import kotlin.reflect.KClass
+
+/** Marks a class as a gRPC call handler. */
+@Target(AnnotationTarget.CLASS) annotation class GrpcServiceHandler(val grpcWrapperType: KClass<*>)
