+++ io_grpc/example/service/armeria/src/main/kotlin/com/geekinasuit/daggergrpc/iogrpc/example/armeria/ExampleServer.kt
@@ -1,5 +1,6 @@
 package com.geekinasuit.daggergrpc.iogrpc.example.armeria
 
+import com.geekinasuit.daggergrpc.armeria.wrapService
 import com.linecorp.armeria.server.Server
 import io.github.oshai.kotlinlogging.KotlinLogging
