+++ io_grpc/example/service/armeria/src/main/kotlin/com/geekinasuit/daggergrpc/iogrpc/example/armeria/WhateverService.kt
@@ -1,15 +1,16 @@
 package com.geekinasuit.daggergrpc.iogrpc.example.armeria
 
-import com.geekinasuit.daggergrpc.iogrpc.example.armeria.api.HelloWorld
-import com.geekinasuit.daggergrpc.iogrpc.example.armeria.api.Whatever
+import com.geekinasuit.daggergrpc.api.GrpcServiceHandler
+import com.geekinasuit.daggergrpc.iogrpc.example.armeria.api.Whatever.WhateverRequest
 import com.geekinasuit.daggergrpc.iogrpc.example.armeria.api.Whatever.WhateverResponse
 import com.geekinasuit.daggergrpc.iogrpc.example.armeria.api.WhateverServiceGrpc
 import io.grpc.stub.StreamObserver
 
+@GrpcServiceHandler(WhateverServiceGrpc::class)
 class WhateverService : WhateverServiceGrpc.AsyncService {
   override fun whatever(
-    request: Whatever.WhateverRequest,
-    responseObserver: StreamObserver<Whatever.WhateverResponse>
+    request: WhateverRequest,
+    responseObserver: StreamObserver<WhateverResponse>
   ) {
     val response = WhateverResponse.newBuilder().setSuceeded(request.whatever).build()
     responseObserver.onNext(response)
