+++ io_grpc/example/service/armeria/src/main/kotlin/com/geekinasuit/daggergrpc/iogrpc/example/armeria/grpc.kt
@@ -1,22 +1,12 @@
 package com.geekinasuit.daggergrpc.iogrpc.example.armeria
 
-import com.linecorp.armeria.server.grpc.GrpcService
-import io.grpc.BindableService
+
 import io.grpc.Metadata
 import io.grpc.ServerCall
 import io.grpc.ServerCallHandler
 import io.grpc.ServerInterceptor
-import kotlin.reflect.KClass
 
-/** Marks a class as a gRPC call handler. */
-@Target(AnnotationTarget.CLASS) annotation class GrpcServiceHandler(val grpcWrapperType: KClass<*>)
 
-fun wrapService(
-  bindableService: BindableService,
-  vararg interceptors: ServerInterceptor,
-): GrpcService {
-  return GrpcService.builder().addService(bindableService).intercept(*interceptors).build()
-}
 
 /**
  * An interceptor, which puts the [GrpcCallContext] (containing http headers [Metadata] and
