+++ util/armeria/src/main/kotlin/com/geekinasuit/daggergrpc/armeria/armeria_grpc_util.kt.kt
@@ -1,0 +1,12 @@
+package com.geekinasuit.daggergrpc.armeria
+
+import com.linecorp.armeria.server.grpc.GrpcService
+import io.grpc.BindableService
+import io.grpc.ServerInterceptor
+
+fun wrapService(
+  bindableService: BindableService,
+  vararg interceptors: ServerInterceptor,
+): GrpcService {
+  return GrpcService.builder().addService(bindableService).intercept(*interceptors).build()
+}
